# doc/api.yaml
openapi: 3.0.0
info: 
  title: WASA Photo
  description:  |
    APIs designed for a social focused on photos
  version: 1.0.0
...
paths:
...
  /users/{userId}/uploadPhoto:
    post:
      tags:
        - User
      summary: Upload photo
      description: Upload a photo for the current user
      security:
        - bearerAuth: []
      operationId: uploadPhoto
      parameters:
        - name: userId
          in: path
          description: Username of current user
          required: true
          schema:
            $ref: '#/components/schemas/Identifier'
      requestBody:
        content:
          image/png:
            schema:
              type: string
              description: String for png representation
              minLength: 1
              maxLength: 500
              format: binary
          image/jpg:
            schema:
              type: string
              description: String for jpg representation
              minLength: 1
              maxLength: 500
              format: binary
          image/jpeg:
            schema:
              type: string
              description: String for jpeg representation
              minLength: 1
              maxLength: 500
              format: binary
      responses:
        '201':
          description: Photo uploaded successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Photo"
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '500':
          description: The server encountered an internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
...
components:
  securitySchemes:

    bearerAuth:
      type: http
      scheme: Bearer
  schemas:
    Identifier:
      readOnly: true
      description: Unique identifier used for several components
      type: integer
      example: 6365 

#    UserIdentifier:
#      description: Unique identifier for users 
#      type: string
#      pattern: "[a-z0-9]+"
#      minLength: 3
#      maxLength: 16
#      example: abcdef012345
...
    Photo:   
      required:
        - photo_id
        - photo_url
        - likes
        - date_time
        - comments
      type: object
      description: Photo of platform
      properties:
        photo_id:
          $ref: '#/components/schemas/Identifier'
        photo_file:
          description: File of the photo
          type: string
          format: binary
          minLength: 1
          maxLength: 99999
        likes:
          type: array
          minItems: 0
          maxItems: 50
          description: List of likes
          items:
            $ref: "#/components/schemas/Like"
        date_time:
          type: string
          minLength: 20
          maxLength: 20
          format: date-time
          description: Date and time of post 
          example: "2017-07-21T17:32:28Z"
        comments:
          type: array
          minItems: 0
          maxItems: 50
          description: List of comments
          items:
            $ref: "#/components/schemas/Comment"
    Comment:  #A comment
      required:
        - comment_id
        - text
        - userId
      type: object
      description: Comment to a photo
      properties:
        comment_id: 
          $ref: '#/components/schemas/Identifier'
        userId:
          $ref: '#/components/schemas/Identifier'
        text:
          type: string
          minLength: 1
          maxLength: 5000
          pattern: '^.*?$'
          description: A simple text
          example: Sample text
    Like:
      required:
        - photoId
        - user_Id
      type: object
      description: Like to a photo
      properties:
        photoId:
          $ref: '#/components/schemas/Identifier'
        user_Id:
          $ref: '#/components/schemas/Identifier'
    ApiResponse:
      type: object
      description: Text of responses
      properties:
        code:
          type: integer
          format: int32
          description: Response code
        type:
          type: string
          minLength: 1
          maxLength: 50
          description: Type of response
        message:
          type: string
          minLength: 1
          maxLength: 5000
          description: Text of response